requirements.txt
streamlit
pandas
numpy
matplotlib
joblib
python-dotenv
google-generativeai
